Split ComponentData implementations
Splits ComponentDataImpl into:
- ComponentDataImpl
- ContainerDataImpl
- ImageDataImpl
- PageDataImpl
