+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/models/datalayer/builder/ContainerDataBuilder.java
@@ -15,7 +15,7 @@
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
 package com.adobe.cq.wcm.core.components.models.datalayer.builder;
 
-import com.adobe.cq.wcm.core.components.internal.models.v1.datalayer.ComponentDataImpl;
+import com.adobe.cq.wcm.core.components.internal.models.v1.datalayer.ContainerDataImpl;
 import com.adobe.cq.wcm.core.components.internal.models.v1.datalayer.builder.DataLayerSupplier;
 import com.adobe.cq.wcm.core.components.internal.models.v1.datalayer.builder.DataLayerSupplierImpl;
 import com.adobe.cq.wcm.core.components.models.datalayer.ContainerData;
@@ -59,6 +59,6 @@ public final class ContainerDataBuilder extends GenericComponentDataBuilder<Cont
     @NotNull
     @Override
     public ContainerData build() {
-        return new ComponentDataImpl(this.getDataLayerSupplier());
+        return new ContainerDataImpl(this.getDataLayerSupplier());
     }
 }
